requirements.txt
GitPython
numpy
tqdm
coloredlogs
jupyter
seaborn
# torch==1.10.0+cu113

# PyTorch pip package
# --find-links https://download.pytorch.org/whl/torch_stable.html
